移除 过期活动手机尚学季脚本
